requirements.txt
dominate==2.6.0                          
ffmpeg-python==0.2.0                            
imageio==2.9.0              
imageio-ffmpeg==0.4.4              
matplotlib==3.3.4                           
numpy==1.18.3             
opencv-python==4.2.0.34            
pandas==1.0.5              
Pillow==8.1.0
scikit-image          
scikit-learn           
scipy 
torch==1.7.1              
torchaudio==0.7.2              
torchfile==0.1.0              
torchvision==0.8.2              
tqdm==4.46.0